+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 1.2.1
 * Enhanced the display of over-time healing
   * Now shows every half hearth like health
   * Bar is now more towards yellow and red when lower over-time healing strength